requirements.txt
azure-ai-vision-imageanalysis
streamlit
pillow
opencv-python